% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pokemon.R
\name{get_pokemon}
\alias{get_pokemon}
\title{Get Pokemon}
\usage{
get_pokemon(x, ...)
}
\arguments{
\item{x}{Pokemon name or ID.}

\item{...}{Any other optional \href{https://pokeapi.co/docs/v2}{parameter} that can be provided.}
}
\value{
A dataframe for the requested Pokemon.
}
\description{
The \code{get_pokemon()} function is used to get structured data from the
\href{https://pokeapi.co/}{"PokeAPI"} API service.
}
\details{
There are two ways to get the data:
\enumerate{
  \item By providing the Pokemon name (lowercase) or Pokemon ID.
  \item By providing \code{limit} and \code{offset} parameters.
}
}
\examples{
\dontrun{
mew <- get_pokemon("mew")
gen_1 <- get_pokemon(limit = 151, offset = 0)
}
}
